Consolidate logic between DistributableCacheFactoryBuilderServiceConfigurator and LegacyDistributableCacheFactoryBuilderServiceConfigurator.
