+++ ejb3/src/main/java/org/jboss/as/ejb3/cache/distributable/DistributableCacheFactoryBuilderServiceConfigurator.java
@@ -21,36 +21,16 @@
  */
 package org.jboss.as.ejb3.cache.distributable;
 
-import java.time.Duration;
-import java.util.concurrent.TimeUnit;
-import java.util.function.Consumer;
-
-import org.jboss.as.clustering.controller.ResourceServiceConfigurator;
 import org.jboss.as.controller.OperationContext;
 import org.jboss.as.controller.OperationFailedException;
-import org.jboss.as.clustering.controller.CapabilityServiceConfigurator;
 import org.jboss.as.controller.PathAddress;
-import org.jboss.as.ee.component.ComponentConfiguration;
-import org.jboss.as.ejb3.cache.CacheFactoryBuilder;
 import org.jboss.as.ejb3.cache.Contextual;
 import org.jboss.as.ejb3.cache.Identifiable;
-import org.jboss.as.ejb3.component.stateful.StatefulComponentDescription;
-import org.jboss.as.ejb3.component.stateful.StatefulTimeoutInfo;
-import org.jboss.as.server.deployment.Attachments;
-import org.jboss.as.server.deployment.DeploymentUnit;
 import org.jboss.as.ejb3.subsystem.DistributableCacheFactoryResourceDefinition;
 import org.jboss.dmr.ModelNode;
-import org.jboss.modules.Module;
-import org.jboss.msc.Service;
-import org.jboss.msc.service.ServiceBuilder;
-import org.jboss.msc.service.ServiceName;
-import org.jboss.msc.service.ServiceTarget;
 import org.wildfly.clustering.ee.Batch;
-import org.wildfly.clustering.ejb.StatefulBeanConfiguration;
-import org.wildfly.clustering.ejb.DistributableBeanManagementProvider;
 import org.wildfly.clustering.ejb.EjbProviderRequirement;
 import org.wildfly.clustering.service.ServiceConfigurator;
-import org.wildfly.clustering.service.SupplierDependency;
 import org.wildfly.clustering.service.ServiceSupplierDependency;
 
 /**
@@ -62,73 +42,17 @@ import org.wildfly.clustering.service.ServiceSupplierDependency;
  * @param <K> the cache key type
  * @param <V> the cache value type
  */
-public class DistributableCacheFactoryBuilderServiceConfigurator<K, V extends Identifiable<K> & Contextual<Batch>> extends DistributableCacheFactoryBuilderServiceNameProvider implements ResourceServiceConfigurator, CacheFactoryBuilder<K, V> {
+public class DistributableCacheFactoryBuilderServiceConfigurator<K, V extends Identifiable<K> & Contextual<Batch>> extends AbstractDistributableCacheFactoryBuilderServiceConfigurator<K, V> {
 
-    private volatile SupplierDependency<DistributableBeanManagementProvider> provider;
-
-    public DistributableCacheFactoryBuilderServiceConfigurator(PathAddress pathAddress) {
-        super(pathAddress.getLastElement().getValue());
+    public DistributableCacheFactoryBuilderServiceConfigurator(PathAddress address) {
+        super(address);
     }
 
     @Override
     public ServiceConfigurator configure(OperationContext context, ModelNode model) throws OperationFailedException {
         // if the attribute is undefined, pass null when generating the service name to pick up the default bean management provider
         String provider = DistributableCacheFactoryResourceDefinition.Attribute.BEAN_MANAGEMENT.resolveModelAttribute(context, model).asStringOrNull();
-        this.provider = new ServiceSupplierDependency<>(EjbProviderRequirement.BEAN_MANAGEMENT_PROVIDER.getServiceName(context, provider));
+        this.accept(new ServiceSupplierDependency<>(EjbProviderRequirement.BEAN_MANAGEMENT_PROVIDER.getServiceName(context, provider)));
         return this;
-    }
-
-    @Override
-    public ServiceBuilder<?> build(ServiceTarget target) {
-        ServiceName name = this.getServiceName();
-        ServiceBuilder<?> builder = target.addService(name);
-        Consumer<CacheFactoryBuilder<K, V>> cacheFactoryBuilder = this.provider.register(builder).provides(name);
-        Service service = Service.newInstance(cacheFactoryBuilder, this);
-        return builder.setInstance(service);
-    }
-
-    @Override
-    public Iterable<CapabilityServiceConfigurator> getDeploymentServiceConfigurators(DeploymentUnit unit) {
-        return this.provider.get().getDeploymentServiceConfigurators(unit.getServiceName());
-    }
-
-    @Override
-    public CapabilityServiceConfigurator getServiceConfigurator(DeploymentUnit unit, StatefulComponentDescription description, ComponentConfiguration configuration) {
-        StatefulBeanConfiguration statefulBeanConfiguration = new StatefulBeanConfiguration() {
-            @Override
-            public String getName() {
-                return configuration.getComponentName();
-            }
-
-            @Override
-            public ServiceName getDeploymentUnitServiceName() {
-                return unit.getServiceName();
-            }
-
-            @Override
-            public Module getModule() {
-                return unit.getAttachment(Attachments.MODULE);
-            }
-
-            @Override
-            public Duration getTimeout() {
-                StatefulTimeoutInfo info = description.getStatefulTimeout();
-
-                // A value of -1 means the bean will never be removed due to timeout
-                if (info == null || info.getValue() < 0) {
-                    return null;
-                }
-                // TODO Once based on JDK9+, change to Duration.of(this.info.getValue(), this.info.getTimeUnit().toChronoUnit())
-                return Duration.ofMillis(TimeUnit.MILLISECONDS.convert(info.getValue(), info.getTimeUnit()));
-            }
-        };
-        CapabilityServiceConfigurator configurator = this.provider.get().getBeanManagerFactoryServiceConfigurator(statefulBeanConfiguration);
-        // name vs description.getCacheFactoryServiceName()
-        return new DistributableCacheFactoryServiceConfigurator<K, V>(description.getCacheFactoryServiceName(), configurator);
-    }
-
-    @Override
-    public boolean supportsPassivation() {
-        return true;
     }
 }
